+++ core/scripts/checkTTY.sh
@@ -10,23 +10,32 @@
 # checkTTY.sh <portName> <zigateType> <gpioLib> [prefix]
 # ex: checkTTY.sh /dev/ttyS1 PI WiringPi "[2021-03-18 12:44:09] "
 
+ABEILLEROOT=${PWD}/../..
 PREFIX=$4
 
 # NOW=`date +"%Y-%m-%d %H:%M:%S"`
 # echo "[${NOW}] Démarrage de '$(basename $0)'"
-echo "${PREFIX}Démarrage de '$(basename $0)'"
+echo "${PREFIX}Test de port"
 
-if [ $# -lt 2 ]; then
-    echo "${PREFIX}= ERREUR: Port et/ou type manquant !"
+if [ $# -lt 3 ]; then
+    echo "${PREFIX}= ERREUR: Argument(s) manquant(s) !"
+    echo "${PREFIX}= Usage: checkTTY.sh <portName> <zigateType> <gpioLib> [prefix]"
     exit 1
 fi
 PORT=$1
 TYPE=$2
 GPIOLIB=$3
-
-echo "${PREFIX}- Vérifications du port '${PORT}'"
+if [ "${TYPE}" != "PI" ] && [ "${TYPE}" != "PIv2" ] && [ "${TYPE}" != "DIN" ]; then
+    echo "${PREFIX}= ERREUR: Type ${TYPE} invalide !"
+    exit 1
+fi
+if [ "${GPIOLIB}" != "WiringPi" ] && [ "${GPIOLIB}" != "PiGpio" ]; then
+    echo "${PREFIX}= ERREUR: Lib GPIO ${GPIOLIB} invalide !"
+    exit 1
+fi
 
 # Port exists ?
+echo "${PREFIX}- Vérifications du port '${PORT}'"
 if [ ! -e ${PORT} ]; then
     echo "${PREFIX}= ERREUR: Le port ${PORT} n'existe pas !"
     exit 2
@@ -45,6 +54,7 @@ else
             break
         fi
     done
+
     echo "${PREFIX}= ERREUR: Le port est utilisé par le process '${PID}'."
     echo "${PREFIX}=         Il doit être libéré et n'être utilisé QUE par le plugin Abeille pour permettre le dialogue avec la Zigate."
     PSOUT=`ps --pid ${PID} -o ppid,cmd | grep -v PPID`
@@ -54,7 +64,18 @@ else
     CMD=${PSOUTA[@]:1}
     echo "${PREFIX}=         Details du process ${PID}:"
     echo "${PREFIX}=           PPid=${PPID2}, cmd='${CMD}'"
-    exit 3
+
+    echo
+    echo "${PREFIX}= Additional infos I"
+    dmesg | grep tty
+
+    echo
+    echo "${PREFIX}= Additional infos II"
+    ls -al /dev/serial*
+
+    # exit 3
+    echo "- Arret forcé du processus ${PID}"
+    kill -9 ${PID}
 fi
 
 # If Zigate type is "PI", let's check if "WiringPi" is properly installed
@@ -82,7 +103,7 @@ if [ "${TYPE}" == "PI" ] || [ "${TYPE}" == "PIv2" ]; then
         fi
         echo "${PREFIX}= Ok"
     elif [ ${GPIOLIB} == "PiGpio" ]; then
-        python /var/www/html/plugins/Abeille/core/scripts/resetPiZigate.py
+        python3 ${ABEILLEROOT}/core/scripts/resetPiZigate.py
     fi
 
     echo "${PREFIX}- Configuration du port série"